Remove market calls for graph and improve production calc
The market calls were causing issues, and I don't care about that data
it was trying to pull so I removed it. I also upgraded the production
tab to allow you to enter in your guzz level and it'll properly calc it,
also now you can use the upgrade items button for all items. It gives
some less than idea results when you do it(as it will assume you want to
make items like holy cheese from holy milk, which you will likely never
want to do), but I'm not sure how to best handle it. So for now, the
best option is to just manually enter in the cost of the item those
things make so you can get the more accurate price.